+++ src/handlers/meta_transaction_handlers.ts
@@ -19,20 +19,20 @@ import {
 import { logger } from '../logger';
 import { schemas } from '../schemas';
 import { MetaTransactionService } from '../services/meta_transaction_service';
-import { MetaTransactionPriceResponse, MetaTransactionQuoteRequestParams } from '../types';
+import { MetaTransactionPriceResponse, MetaTransactionQuoteRequestParams, IMetaTransactionService } from '../types';
 import { findTokenAddressOrThrowApiError } from '../utils/address_utils';
 import { parseUtils } from '../utils/parse_utils';
 import { schemaUtils } from '../utils/schema_utils';
 
 export class MetaTransactionHandlers {
-    private readonly _metaTransactionService: MetaTransactionService;
+    private readonly _metaTransactionService: IMetaTransactionService;
 
     public static rootAsync(_req: express.Request, res: express.Response): void {
         const message = `This is the root of the Meta Transaction API. Visit ${META_TRANSACTION_DOCS_URL} for details about this API.`;
         res.status(StatusCodes.OK).send({ message });
     }
 
-    constructor(metaTransactionService: MetaTransactionService) {
+    constructor(metaTransactionService: IMetaTransactionService) {
         this._metaTransactionService = metaTransactionService;
     }
