+++ src/routers/meta_transaction_router.ts
@@ -1,10 +1,11 @@
 import * as express from 'express';
 import * as asyncHandler from 'express-async-handler';
 
+import { IMetaTransactionService } from '../types';
 import { MetaTransactionHandlers } from '../handlers/meta_transaction_handlers';
 import { MetaTransactionService } from '../services/meta_transaction_service';
 
-export const createMetaTransactionRouter = (metaTransactionService: MetaTransactionService): express.Router => {
+export const createMetaTransactionRouter = (metaTransactionService: IMetaTransactionService): express.Router => {
     const router = express.Router();
     const handlers = new MetaTransactionHandlers(metaTransactionService);
