+++ src/runners/http_service_runner.ts
@@ -3,7 +3,7 @@ import * as express from 'express';
 import * as core from 'express-serve-static-core';
 import { Server } from 'http';
 
-import { AppDependencies, getDefaultAppDependenciesAsync } from '../app';
+import { getDefaultAppDependenciesAsync } from '../app';
 import {
     defaultHttpServiceConfig,
     SENTRY_DSN,
@@ -22,7 +22,7 @@ import { createSRARouter } from '../routers/sra_router';
 import { createSwapRouter } from '../routers/swap_router';
 import { SentryInit, SentryOptions } from '../sentry';
 import { WebsocketService } from '../services/websocket_service';
-import { HttpServiceConfig } from '../types';
+import { HttpServiceConfig, AppDependencies } from '../types';
 import { providerUtils } from '../utils/provider_utils';
 
 import { destroyCallback } from './utils';
